requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
streamlit>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
requests>=2.26.0 